packages/main/pyproject.toml
[tool.poetry]
name = "rpaframework"
version = "5.1.0"
description = "A collection of tools and libraries for RPA"
authors = [
	"RPA Framework <rpafw@robocorp.com>",
]
license = "Apache-2.0"
readme = "README.rst"

homepage = "https://rpaframework.org/"
documentation = "https://rpaframework.org/"
repository = "https://github.com/robocorp/rpaframework"

keywords = ["robotframework", "rpa", "automation"]
classifiers = [
	"License :: OSI Approved :: Apache Software License",
	"Development Status :: 3 - Alpha",
	"Operating System :: OS Independent",
	"Intended Audience :: Developers",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Software Development :: Libraries",
	"Framework :: Robot Framework :: Library",
	"Framework :: Robot Framework",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
]

packages = [
	{ include = "RPA", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.6"
dataclasses = { version = "^0.7", python=">=3.6,<3.7" }
rpaframework-core = { path = "../core/" }
rpaframework-recognition = { path = "../recognition", optional = true }
robotframework = "^3.2.2"
robotframework-sapguilibrary = { version = "^1.1", platform = "win32" }
robotframework-seleniumlibrary = "^4.5.0"
robotframework-seleniumtestability = "^1.1.0"
robotframework-requests = "^0.6.5"
pywinauto = { version = "^0.6.8", platform = "win32", python = "!=3.8.1,!=3.7.6" }
pywin32 = { version = "^227", platform = "win32", python = "!=3.8.1,!=3.7.6" }
robotframework-pythonlibcore = "^2.1.0"
pynput = "^1.7.0"
evdev = { version = ">=1.3", platform = "linux" }
python-xlib = { version = ">=0.17", platform = "linux" }
psutil = { version = "^5.7.0", platform = "win32" }
pyperclip = "^1.8.0"
exchangelib = "^3.1.1"
fpdf = "^1.7.2"
xlrd = "^1.2.0"
xlwt = "^1.3.0"
xlutils = "^2.0.0"
openpyxl = "^3.0.3"
pillow = "^7.0.0"
simple_salesforce = "^1.0.0"
netsuitesdk = "^1.1.0"
"pdfminer.six" = "^20200402"
pypdf2 = "^1.26.0"
tweepy = "^3.8.0"
boto3 = { version = "^1.13.4", optional = true }
google-api-python-client = { version = "^1.12.3", optional = true }
google-auth-httplib2 = { version = "^0.0.4", optional = true }
google-auth-oauthlib = { version = "^0.4.1", optional = true }
google-cloud-language = { version = "^1.3.0", optional = true }
google-cloud-speech = { version = "^1.3.2", optional = true }
google-cloud-storage = { version = "^1.28.1", optional = true }
google-cloud-texttospeech = { version = "^1.0.1", optional = true }
google-cloud-translate = { version = "^2.0.1", optional = true }
google-cloud-videointelligence = { version = "^1.14.0", optional = true }
google-cloud-vision = { version = "^1.0.0", optional = true }
grpcio = { version = "^1.29.0", optional = true }
graphviz = "^0.13.2"
notifiers = "^1.2.1"
cryptography = "^2.9.2"
mss = "^6.0.0"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
flake8 = "^3.7.9"
pylint = "^2.4.4"
pytest = "^5.2"
mock = "^4.0.2"
docutils = "^0.14"
mypy = "^0.782"

[tool.poetry.extras]
cv = ["rpaframework-recognition"]
aws = ["boto3"]
google = [
	"google-api-python-client",
	"google-auth-httplib2",
	"google-auth-oauthlib",
	"google-cloud-language",
	"google-cloud-speech",
	"google-cloud-storage",
	"google-cloud-texttospeech",
	"google-cloud-translate",
	"google-cloud-videointelligence",
	"google-cloud-vision",
	"grpcio"
]

[tool.black]
target-version = ["py36", "py37", "py38"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
